refactor: move CalendarPicker components to NewDatePicker dir
