+++ src/components/NewDatePicker/CalendarPicker/ArrowIcon.js
@@ -1,11 +1,11 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import {View} from 'react-native';
-import PropTypes from 'prop-types';
-import styles from '../../styles/styles';
-import * as Expensicons from '../Icon/Expensicons';
-import * as StyleUtils from '../../styles/StyleUtils';
-import Icon from '../Icon';
-import CONST from '../../CONST';
+import CONST from '../../../CONST';
+import * as StyleUtils from '../../../styles/StyleUtils';
+import styles from '../../../styles/styles';
+import Icon from '../../Icon';
+import * as Expensicons from '../../Icon/Expensicons';
 
 const propTypes = {
     /** Specifies if the arrow icon should be disabled or not. */
