+++ src/components/NewDatePicker/CalendarPicker/YearPickerPage.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import _ from 'underscore';
-import CONST from '../../CONST';
-import styles from '../../styles/styles';
-import themeColors from '../../styles/themes/default';
-import HeaderWithBackButton from '../HeaderWithBackButton';
-import * as Expensicons from '../Icon/Expensicons';
-import OptionsSelector from '../OptionsSelector';
-import ScreenWrapper from '../ScreenWrapper';
-import {withCurrentUserPersonalDetailsDefaultProps, withCurrentUserPersonalDetailsPropTypes} from '../withCurrentUserPersonalDetails';
-import withLocalize, {withLocalizePropTypes} from '../withLocalize';
+import CONST from '../../../CONST';
+import styles from '../../../styles/styles';
+import themeColors from '../../../styles/themes/default';
+import HeaderWithBackButton from '../../HeaderWithBackButton';
+import * as Expensicons from '../../Icon/Expensicons';
+import OptionsSelector from '../../OptionsSelector';
+import ScreenWrapper from '../../ScreenWrapper';
+import {withCurrentUserPersonalDetailsDefaultProps, withCurrentUserPersonalDetailsPropTypes} from '../../withCurrentUserPersonalDetails';
+import withLocalize, {withLocalizePropTypes} from '../../withLocalize';
 
 const greenCheckmark = {src: Expensicons.Checkmark, color: themeColors.success};
