+++ src/components/NewDatePicker/CalendarPicker/index.js
@@ -5,14 +5,14 @@ import React from 'react';
 import {StyleSheet, View} from 'react-native';
 import Animated, {SlideInRight, SlideOutLeft} from 'react-native-reanimated';
 import _ from 'underscore';
-import CONST from '../../CONST';
-import getButtonState from '../../libs/getButtonState';
-import * as StyleUtils from '../../styles/StyleUtils';
-import styles from '../../styles/styles';
-import PressableWithFeedback from '../Pressable/PressableWithFeedback';
-import PressableWithoutFeedback from '../Pressable/PressableWithoutFeedback';
-import Text from '../Text';
-import withLocalize from '../withLocalize';
+import CONST from '../../../CONST';
+import getButtonState from '../../../libs/getButtonState';
+import * as StyleUtils from '../../../styles/StyleUtils';
+import styles from '../../../styles/styles';
+import PressableWithFeedback from '../../Pressable/PressableWithFeedback';
+import PressableWithoutFeedback from '../../Pressable/PressableWithoutFeedback';
+import Text from '../../Text';
+import withLocalize from '../../withLocalize';
 import ArrowIcon from './ArrowIcon';
 import YearPickerPage from './YearPickerPage';
 import {propTypes as calendarPickerPropType, defaultProps as defaultCalendarPickerPropType} from './calendarPickerPropTypes';
